move adaptive wingloss to heatmap_loss
