+++ mmpose/models/losses/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .classfication_loss import BCELoss
+from .heatmap_loss import AdaptiveWingLoss
 from .mesh_loss import GANLoss, MeshLoss
 from .mse_loss import JointsMSELoss, JointsOHKMMSELoss
 from .multi_loss_factory import AELoss, HeatmapLoss, MultiLossFactory
