+++ tests/test_losses/test_top_down_losses.py
@@ -2,10 +2,33 @@
 import pytest
 import torch
 
+from mmpose.models import build_loss
+
+
+def test_adaptive_wing_loss():
+    # test Adaptive WingLoss without target weight
+    loss_cfg = dict(type='AdaptiveWingLoss')
+    loss = build_loss(loss_cfg)
+
+    fake_pred = torch.zeros((1, 3, 64, 64))
+    fake_label = torch.zeros((1, 3, 64, 64))
+    assert torch.allclose(loss(fake_pred, fake_label, None), torch.tensor(0.))
+
+    # test WingLoss with target weight
+    loss_cfg = dict(type='AdaptiveWingLoss', use_target_weight=True)
+    loss = build_loss(loss_cfg)
+
+    fake_pred = torch.ones((1, 3, 64, 64))
+    fake_label = torch.zeros((1, 3, 64, 64))
+    assert torch.allclose(
+        loss(fake_pred, fake_label, torch.ones_like(fake_label)),
+        torch.tensor(0.))
+
+
+test_adaptive_wing_loss()
+
 
 def test_mse_loss():
-    from mmpose.models import build_loss
-
     # test MSE loss without target weight
     loss_cfg = dict(type='JointsMSELoss')
     loss = build_loss(loss_cfg)
@@ -70,8 +93,6 @@ def test_mse_loss():
 
 
 def test_smoothl1_loss():
-    from mmpose.models import build_loss
-
     # test MSE loss without target weight
     loss_cfg = dict(type='SmoothL1Loss')
     loss = build_loss(loss_cfg)
